+++ components/mypage/content/Workplace.module.css
@@ -1,0 +1,57 @@
+.container {
+    width: 100%;
+}
+
+.otherComponent {
+    margin: 2rem 0;
+}
+
+.addButtonContainer {
+    display: flex;
+    justify-content: center;
+    margin-top: 1.5rem;
+}
+
+.addButton {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    color: #9ca3af;  /* gray-400 */
+    transition: color 0.2s;
+}
+
+.addButton:hover {
+    color: #4b5563;  /* gray-600 */
+}
+
+.iconContainer {
+    width: 3rem;
+    height: 3rem;
+    border-radius: 9999px;
+    border: 2px solid currentColor;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 0.5rem;
+}
+
+.icon {
+    width: 1.5rem;
+    height: 1.5rem;
+}
+
+.addButtonText {
+    font-size: 1.125rem;
+}
+
+.accountContainer {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 5px;
+}
+
+.bankLogo {
+    width: 20px;
+    height: 20px;
+}
